Relocate /dev to /admin/dev and enforce developer role check
