+++ app/frontend/src/router/index.js
@@ -25,20 +25,17 @@ const router = new VueRouter({
           component: () => import(/* webpackChunkName: "submission-table" */ '@/views/admin/Root.vue')
         },
         {
+          path: 'dev',
+          name: 'Dev',
+          component: () => import(/* webpackChunkName: "dev" */ '@/views/admin/Dev.vue'),
+        },
+        {
           path: 'submission/:ipcPlanId',
           name: 'Submission',
           component: () => import(/* webpackChunkName: "submission" */ '@/views/admin/Submission.vue'),
           props: true
         }
       ],
-      meta: {
-        requiresAuth: true
-      }
-    },
-    {
-      path: '/dev',
-      name: 'Dev',
-      component: () => import(/* webpackChunkName: "dev" */ '@/views/Dev.vue'),
       meta: {
         requiresAuth: true
       }
